Second Try New/Continue Game Fixed
Previously when a player would go back to the main menu through the pause menu, the player would be unable to continue the game and would be stuck upon trying to create a new game. This issue should be fixed and the player should be able to start a new game or continue their game from the main menu and properly restart the game from the win screen.